move ContextExecutor to own package
